Remove ifLoggedIn HOC
This HOC is unnecessary since no scene should ever be rendered which
requires a login state when the user is not logged in. Instead the scene
should crash in some unexpected way rather then show a loading scene
component.

The motivation behind removing this is to simplify our code base and
review crash bugs instead of hide any issues that may arise from
incorrectly rendering scenes.